fix some ads not being recognized
Multiple elements with the same id are being added by google. This wasn't expected as ids are supposed to be unique for each element. Before they had an id for top and bottom ad sections. Now that they switched to endless scroll they are adding multiple ad units to the same page. I assume that they are adding the same component to the page again and again but they haven't removed the id from it.